Change LLA from float to double in VN200Data.
Changes the data type of the latitude, longitude, and altitude messages
in the VN200Data structure from float to double so as not to lose
precision. The VN-200 sensor reports these as floats, but in the
previous version of the VN200Data class these were truncated to single
precision floating points, but then logged with the precision expected
from a double precision float.

+++ src/vn200/vn200.hh
@@ -68,7 +68,7 @@ struct VN200Data
      *
      *  Order is lat, long, alt.
      */
-    float posDegDegM[3] = {0};
+    double posDegDegM[3] = {0};
     //! NED Velocity
     /*!
      *  Stores the estimated velocity in the North-East-Down frame in m/s.
